plugins: Switch to using a uint32 for versioning
This allows us to do semantic versioning without any strings.

+++ internal/plugins/plugins.go
@@ -34,7 +34,7 @@ type Plugin interface {
 
 	FilePath() string
 
-	Version() uint16
+	Version() uint32
 
 	Description() string
